requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djhtml==3.0.6
filelock==3.12.2
greenlet==2.0.2
msgpack==1.0.5
platformdirs==3.10.0
pynvim==0.4.3
pytz==2023.3
sqlparse==0.4.4
virtualenv==20.24.2
